src/third_party/blink/web_tests/external/wpt/workers/constructors/SharedWorker/undefined-arguments-expected.txt

This is a testharness.js-based test.
FAIL undefined as arguments assert_equals: second arg expected "undefined" but got ""
Harness: the test ran to completion.
